+++ app/models/variable.rb
@@ -24,7 +24,7 @@ class Variable < ActiveRecord::Base
   belongs_to :domain
   has_many :tags, dependent: :destroy
   has_many :mappings, dependent: :destroy
-  has_many :query_concepts, dependent: :destroy
+  has_many :criteria, dependent: :destroy
   has_many :report_concepts, dependent: :destroy
   has_many :sources, -> { where(deleted: false).uniq.order('sources.name') }, through: :mappings