+++ db/migrate/20130722180712_rename_query_concepts_to_criteria.rb
@@ -1,0 +1,5 @@
+class RenameQueryConceptsToCriteria < ActiveRecord::Migration
+  def change
+    rename_table :query_concepts, :criteria
+  end
+end