+++ test/helpers/query_concepts_helper_test.rb
@@ -1,4 +1,0 @@
-require 'test_helper'
-
-class QueryConceptsHelperTest < ActionView::TestCase
-end