+++ test/models/criterium_test.rb
@@ -1,0 +1,5 @@
+require 'test_helper'
+
+class CriteriumTest < ActiveSupport::TestCase
+
+end